Do not inject mixins into base objects dynamically.
