+++ src/ggrc/models/account_balance.py
@@ -28,10 +28,10 @@ class AccountBalance(
     ScopedCommentable,
     mixins.TestPlanned,
     mixins.LastDeprecatedTimeboxed,
+    mixins.CycleTaskable,
     mixins.base.ContextRBAC,
     mixins.ScopeObject,
     mixins.Folderable,
-    mixins.WithWorkflowState,
     fulltext_mixin.Indexed,
     db.Model,
 ):
