+++ src/ggrc/models/control.py
@@ -39,9 +39,9 @@ class Control(with_external_created_by.WithExternalCreatedBy,
               mixins.TestPlanned,
               comment.ExternalCommentable,
               WithSimilarityScore,
+              mixins.CycleTaskable,
               base.ContextRBAC,
               mixins.BusinessObject,
-              mixins.WithWorkflowState,
               Indexed,
               mixins.Folderable,
               db.Model):
