+++ src/ggrc/models/facility.py
@@ -22,10 +22,10 @@ class Facility(Roleable,
                ScopedCommentable,
                mixins.TestPlanned,
                mixins.LastDeprecatedTimeboxed,
+               mixins.CycleTaskable,
                mixins.base.ContextRBAC,
                mixins.ScopeObject,
                mixins.Folderable,
-               mixins.WithWorkflowState,
                Indexed,
                db.Model):
   """Facility model"""
