+++ src/ggrc/models/key_report.py
@@ -25,10 +25,10 @@ class KeyReport(synchronizable.Synchronizable,
                 ScopedCommentable,
                 mixins.TestPlanned,
                 mixins.LastDeprecatedTimeboxed,
+                mixins.CycleTaskable,
                 mixins.base.ContextRBAC,
                 mixins.ScopeObject,
                 mixins.Folderable,
-                mixins.WithWorkflowState,
                 fulltext_mixin.Indexed,
                 db.Model,
                 ):
