+++ src/ggrc/models/market.py
@@ -22,10 +22,10 @@ class Market(Roleable,
              PublicDocumentable,
              ScopedCommentable,
              mixins.TestPlanned,
+             mixins.CycleTaskable,
              mixins.base.ContextRBAC,
              mixins.ScopeObject,
              mixins.Folderable,
-             mixins.WithWorkflowState,
              Indexed,
              db.Model):
   """Market model"""
