+++ src/ggrc/models/mixins/__init__.py
@@ -38,6 +38,7 @@ from ggrc.models.mixins.customattributable import CustomAttributable
 from ggrc.models.mixins.notifiable import Notifiable
 from ggrc.models.mixins.with_workflow_state import WithWorkflowState
 from ggrc.models.mixins.with_external_created_by import WithExternalCreatedBy
+from ggrc.models.mixins.with_workflow_state import CycleTaskable
 from ggrc.models.utils import validate_option
 from ggrc.fulltext import attributes
 
@@ -895,4 +896,5 @@ __all__ = [
     "WithContact",
     "Folderable",
     "WithWorkflowState",
+    "CycleTaskable",
 ]
