+++ src/ggrc/models/product.py
@@ -29,10 +29,10 @@ class Product(Roleable,
               PublicDocumentable,
               ScopedCommentable,
               mixins.TestPlanned,
+              mixins.CycleTaskable,
               mixins.base.ContextRBAC,
               mixins.ScopeObject,
               mixins.Folderable,
-              mixins.WithWorkflowState,
               Indexed,
               db.Model):
   """Representation for Product model."""
