+++ src/ggrc/models/project.py
@@ -21,10 +21,10 @@ class Project(Roleable,
               PublicDocumentable,
               ScopedCommentable,
               mixins.TestPlanned,
+              mixins.CycleTaskable,
               mixins.base.ContextRBAC,
               mixins.ScopeObject,
               mixins.Folderable,
-              mixins.WithWorkflowState,
               Indexed,
               db.Model):
   """Representation for Project model."""
