+++ src/ggrc/models/technology_environment.py
@@ -23,10 +23,10 @@ class TechnologyEnvironment(mixins.CustomAttributable,
                             ScopedCommentable,
                             mixins.TestPlanned,
                             mixins.LastDeprecatedTimeboxed,
+                            mixins.CycleTaskable,
                             mixins.base.ContextRBAC,
                             mixins.ScopeObject,
                             mixins.Folderable,
-                            mixins.WithWorkflowState,
                             db.Model,
                             Indexed):
   """Representation for TechnologyEnvironment model."""
