+++ src/ggrc_workflows/__init__.py
@@ -49,39 +49,7 @@ blueprint = Blueprint(
 )
 
 
-_INJECTABLE_MIXINS = (
-    models.cycle_task_group_object_task.CycleTaskable,
-)
-
-
-def _inject_workflow_mixins():
-  """Inject Workflow mixins to configured model classes"""
-
-  # This function contains slightly refactored code which injects mixins
-  # Mixin injection produces bad side effects, for now only one of them
-  # is found and workarounded.
-  # This code needs to be completely refactored!
-
-  for type_ in WORKFLOW_OBJECT_TYPES:
-    model = getattr(all_models, type_)
-    model.__bases__ = _INJECTABLE_MIXINS + model.__bases__
-    _workaround_mixin_injection(model)
-
-
-def _workaround_mixin_injection(model):
-  """Workaround mixin injection
-
-  We inject mixin to existing Model, and mapper for this Model is
-  already configured. So, all sqlalchemy attrs will not be added to mapper
-  To resolve it, we need to add all attrs to mapper manually.
-  """
-
-  for mixin in _INJECTABLE_MIXINS:
-    for attr_name in getattr(mixin, '_mapper_inject_properties', list()):
-      model.__mapper__.add_property(attr_name, getattr(model, attr_name))
-
-
-def get_public_config(current_user):  # pylint:disable=unused-argument
+def get_public_config(current_user):  # noqa
   """Expose additional permissions-dependent config to client.
   """
   return {}
@@ -949,6 +917,3 @@ contributed_column_handlers = COLUMN_HANDLERS
 contributed_get_ids_related_to = relationship_helper.get_ids_related_to
 NIGHTLY_CRON_JOBS = [start_recurring_cycles]
 NOTIFICATION_LISTENERS = [notification.register_listeners]
-
-
-_inject_workflow_mixins()
