+++ src/ggrc_workflows/models/cycle_task_group_object_task.py
@@ -8,7 +8,6 @@ from logging import getLogger
 from sqlalchemy import orm
 import sqlalchemy as sa
 from sqlalchemy.ext import hybrid
-from sqlalchemy.ext.declarative import declared_attr
 from werkzeug.exceptions import BadRequest
 
 from ggrc import builder
@@ -470,48 +469,3 @@ class CycleTaskGroupObjectTask(roleable.Roleable,
       obj.status = new_state
       obj.modified_by_id = login.get_current_user_id()
     return updated_objects, update_forbidden_objects
-
-
-class CycleTaskable(object):
-  """CycleTaskable mixin"""
-
-  # The following attribute will be used to add sqlalchemy properties
-  # to mapper after Mixin is injected to existing Model class.
-  # Do not forget add new sqlalchemy attrs to the list. These aatrs will be
-  # added to mapper after mixin injection
-  _mapper_inject_properties = [
-      'cycle_task_group_object_tasks',
-  ]
-
-  @declared_attr
-  def cycle_task_group_object_tasks(cls):  # pylint: disable=no-self-argument
-    return db.relationship(
-        CycleTaskGroupObjectTask,
-        primaryjoin=lambda: sa.or_(
-            sa.and_(
-                cls.id == relationship.Relationship.source_id,
-                relationship.Relationship.source_type == cls.__name__,
-                relationship.Relationship.destination_type ==
-                "CycleTaskGroupObjectTask",
-            ),
-            sa.and_(
-                cls.id == relationship.Relationship.destination_id,
-                relationship.Relationship.destination_type == cls.__name__,
-                relationship.Relationship.source_type ==
-                "CycleTaskGroupObjectTask",
-            )
-        ),
-        secondary=relationship.Relationship.__table__,
-        secondaryjoin=lambda: CycleTaskGroupObjectTask.id == sa.case(
-            [(relationship.Relationship.source_type == cls.__name__,
-              relationship.Relationship.destination_id)],
-            else_=relationship.Relationship.source_id
-        ),
-        viewonly=True
-    )
-
-  @classmethod
-  def eager_query(cls, **kwargs):
-    """Eager query for objects with cycle tasks."""
-    query = super(CycleTaskable, cls).eager_query(**kwargs)
-    return query.options(orm.subqueryload('cycle_task_group_object_tasks'))
